backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-multipart>=0.0.9
pandas>=2.0.0
numpy>=1.26.0
yfinance>=0.2.40
alpaca-py>=0.20.0
optuna>=3.6.0
jesse>=0.47.0
psycopg2-binary>=2.9.9
redis>=5.0.0
